+++ SonglistPainter.h
@@ -31,8 +31,8 @@ private:
     int getTableSubTitleBeginY( QRect const & aPrinter ) const;
     int getTableBeginY( QRect const & aPrinter ) const;
 
-    int getHlineBeginY( QPrinter & aPrinter, int aRowIndex ) const;
-    int getRowHeight( QPrinter & aPrinter ) const;
+    double getHlineBeginY( QPrinter & aPrinter, int aRowIndex ) const;
+    double getRowHeight( QPrinter & aPrinter ) const;
     double getColWidth( QPrinter & aPrinter, int aColIndex ) const;
     void drawSongText( QPainter & aPainter, QRect const & aRect, QString const & aString, ColumnType aColType );
     void drawOnePage( QPrinter & aPrinter, QPainter & aPainter, int aSongBeginIndex );
